src/test/kotlin/com/clouddjr/advent2024/Day14Test.kt
package com.clouddjr.advent2024

import com.clouddjr.advent2024.Resources.resourceAsListOfString
import org.junit.jupiter.api.DisplayName
import org.junit.jupiter.api.Nested
import org.junit.jupiter.api.Test
import kotlin.test.assertEquals

@DisplayName("Day 14")
class Day14Test {
    @Nested
    @DisplayName("Part 1")
    inner class Part1 {
        @Test
        fun `Matches actual`() {
            val answer = Day14(resourceAsListOfString("day14.txt")).solvePart1()

            assertEquals(230_900_224, answer)
        }
    }

    @Nested
    @DisplayName("Part 2")
    inner class Part2 {
        @Test
        fun `Matches actual`() {
            val answer = Day14(resourceAsListOfString("day14.txt")).solvePart2()

            assertEquals(6532, answer)
        }
    }
}